+++ koin-projects/koin-core/src/test/java/org/koin/dsl/PropertyDefinitionTest.kt
@@ -15,7 +15,7 @@ class PropertyDefinitionTest {
             properties(values)
         }.koin
 
-        val gotValue = koin.getProperty(key)
+        val gotValue = koin.getProperty<String>(key)
 
         assertEquals(value, gotValue)
     }
@@ -38,7 +38,7 @@ class PropertyDefinitionTest {
         val koin = koinApplication { }.koin
 
         koin.setProperty(key, value)
-        val gotValue = koin.getProperty(key)
+        val gotValue = koin.getProperty<String>(key)
         assertEquals(value, gotValue)
     }
 
@@ -47,7 +47,7 @@ class PropertyDefinitionTest {
         val key = "KEY"
         val koin = koinApplication { }.koin
 
-        val gotValue = koin.getProperty(key)
+        val gotValue = koin.getProperty<String>(key)
         assertNull(gotValue)
     }
 
@@ -62,7 +62,7 @@ class PropertyDefinitionTest {
         }.koin
 
         koin.setProperty(key, value2)
-        val gotValue = koin.getProperty(key)
+        val gotValue = koin.getProperty<String>(key)
         assertEquals(value2, gotValue)
     }
 }